pulls out signing flow account selector handling into shared helper
